requirements.txt
blinker==1.8.2
click==8.1.7
coverage==7.6.0
dateparser==1.2.0
ecdsa==0.19.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycodestyle==2.12.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
pytz==2024.1blinker==1.8.2
click==8.1.7
coverage==7.6.0
dateparser==1.2.0
ecdsa==0.19.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycodestyle==2.12.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
pytz==2024.1
regex==2024.5.15
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzlocal==5.2
Werkzeug==3.0.3
regex==2024.5.15
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzlocal==5.2
Werkzeug==3.0.3